pyproject.toml
[tool.poetry]
name = "polyachatmanager"
version = "0.1.0"
description = ""
authors = ["jotty <bard143games@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^3.0.0a18"
python-dotenv = "*"
tortoise-orm = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
